requirements.txt
beautifulsoup4==4.10.0
certifi==2021.5.30
charset-normalizer==2.0.6
click==8.0.1
colorama==0.4.4
cssselect==1.1.0
feedfinder2==0.0.4
feedparser==6.0.8
filelock==3.1.0
Flask==2.0.1
Flask-Cors==3.0.10
gunicorn==20.1.0
idna==3.2
itsdangerous==2.0.1
jieba3k==0.35.1
Jinja2==3.0.1
joblib==1.0.1
lxml==4.6.3
MarkupSafe==2.0.1
newspaper3k==0.2.8
nltk==3.6.3
numpy==1.21.2
Pillow==8.3.2
python-dateutil==2.8.2
PyYAML==5.4.1
regex==2021.9.24
requests==2.26.0
requests-file==1.5.1
scikit-learn==1.0
scipy==1.7.1
sgmllib3k==1.0.0
six==1.16.0
sklearn==0.0
soupsieve==2.2.1
threadpoolctl==2.2.0
tinysegmenter==0.3
tldextract==3.1.2
tqdm==4.62.3
urllib3==1.26.7
Werkzeug==2.0.1
